Issue #1160: Add a REST-full API for accessing patient data Moved a few classes in a different package
