+++ components/patient-data/rest/src/main/java/org/phenotips/data/rest/DateTimeAdapter.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see http://www.gnu.org/licenses/
  */
-package org.phenotips.data.rest.model;
+package org.phenotips.data.rest;
 
 import javax.xml.bind.annotation.adapters.XmlAdapter;
