+++ components/patient-data/rest/src/main/resources/simple.xjb
@@ -32,6 +32,6 @@
     <xjc:simple/>
 
     <!-- Use JodaTime's DateTime for representing dates instead of JAXB's default, XMLGregorianCalendar. -->
-    <xjc:javaType name="org.joda.time.DateTime" xmlType="xs:dateTime" adapter="org.phenotips.data.rest.model.DateTimeAdapter"/>
+    <xjc:javaType name="org.joda.time.DateTime" xmlType="xs:dateTime" adapter="org.phenotips.data.rest.DateTimeAdapter"/>
   </jaxb:globalBindings>
 </jaxb:bindings>
